# Parallelize some startup on startBitcoinSBackend() to increase performance