This shaves `1 second` on average off of the startup time in #4210

#### Before
>2022-03-23T18:53:08UTC INFO [BitcoinSServerMain] Done starting Main! It took 5015ms

#### After

>2022-03-24T17:58:31UTC INFO [BitcoinSServerMain] Done starting Main! It took 3615ms

>2022-03-24T18:00:52UTC INFO [BitcoinSServerMain] Done starting Main! It took 3660ms

>2022-03-24T18:01:18UTC INFO [BitcoinSServerMain] Done starting Main! It took 3463ms